feat(feed): delete class topic files (replaced by new files), delete ClassFeedList, change editable prop to role prop
